moved useFindPaginated to own file
