+++ composables/api.ts
@@ -1,4 +1,4 @@
-import { keepPreviousData, useQuery } from '@tanstack/vue-query';
+import { useQuery } from '@tanstack/vue-query';
 import axios from 'axios';
 
 export const API_ENDPOINTS = {
@@ -111,98 +111,6 @@ export const useFindMany = (
         findMany(unref(endpoint), unref(sourcesForAPIVersion), unref(params))
       ),
   });
-};
-
-export const useFindPaginated = (options: {
-  endpoint: MaybeRef<string>;
-  itemsPerPage?: MaybeRef<number>;
-  initialPage?: MaybeRef<number>;
-  sortByProperty?: MaybeRef<string>;
-  isSortDescending?: MaybeRef<boolean>;
-  filter?: MaybeRef<Record<string, any>>;
-  params?: MaybeRef<Record<string, any>>;
-}) => {
-  const {
-    endpoint,
-    itemsPerPage = 50,
-    initialPage = 1,
-    sortByProperty = 'name',
-    isSortDescending = false,
-    filter = {},
-    params = {},
-  } = options;
-  const pageNo = ref(unref(initialPage));
-  const { findPaginated } = useAPI();
-  const { sources, sourcesAPIVersion1 } = useSourcesList();
-  const sourcesForAPIVersion = isV1Endpoint(unref(endpoint))
-    ? sourcesAPIVersion1
-    : sources;
-
-  const { data, isFetching, error } = useQuery({
-    queryKey: [
-      'findPaginated',
-      endpoint,
-      sourcesForAPIVersion,
-      itemsPerPage,
-      pageNo,
-      sortByProperty,
-      isSortDescending,
-      filter,
-      params,
-    ],
-    placeholderData: keepPreviousData,
-    queryFn: () =>
-      findPaginated({
-        endpoint: unref(endpoint),
-        sources: unref(sourcesForAPIVersion),
-        pageNo: unref(pageNo),
-        sortByProperty: unref(sortByProperty),
-        isSortDescending: unref(isSortDescending),
-        itemsPerPage: unref(itemsPerPage),
-        queryParams: { ...unref(params), ...unref(filter) },
-      }),
-  });
-
-  const lastPageNo = computed(() => {
-    return data.value ? Math.ceil(data.value.count / unref(itemsPerPage)) : 1;
-  });
-
-  const nextPage = () => {
-    pageNo.value++;
-  };
-
-  const prevPage = () => {
-    if (pageNo.value > 1) {
-      pageNo.value--;
-    }
-  };
-
-  const firstPage = () => {
-    pageNo.value = 1;
-  };
-
-  const lastPage = () => {
-    if (data.value) {
-      pageNo.value = lastPageNo.value;
-    }
-  };
-
-  // Move to the first page when the filter changes, to avoid showing an empty page due to fewer results
-  watch(filter, () => {
-    firstPage();
-  });
-
-  return {
-    data,
-    isFetching,
-    error,
-    firstPage,
-    prevPage,
-    nextPage,
-    lastPage,
-    pageNo,
-    lastPageNo,
-  };
 };
 
 /**
